feat: Update dependencies, refactor validation classes, and add new functions for date handling
