+++ src/main/java/function_fun/VoidNullValidador/FunValStr.java
@@ -1,4 +1,4 @@
-package function_fun.VoudNullValidador;
+package function_fun.VoidNullValidador;
 
 /************************************************************************************************************
  * @author Oliver Consterla Araya                                                                           *
